updating to use a separate lambda for the change of address
